// src/components/forms/RegisterForm/types.tsx
export interface IUserProfileEntity1 {
    email: string;
    phone: string;
    firstName: string;
    lastName: string;
    githubUsername: string;
    portfolioUrls: string[];
    projectUrls: string[];
    bio: string;
    expectedTypeWork: string;
    targetWorkCity: string;
    expectedContractType: string;
    expectedSalary: string;
    canTakeApprenticeship: string;
    monthsOfCommercialExp: number;
    education: string;
    workExperience: string;
    courses: string;
}